xive: Wrap irq_sources to provide standard set_xive/get_xive
All the source controllers use the standard XIVE mechanism for
masking/unmasking and EOI, so there is no point having that
logic duplicated. There are a few variations on how they are
implemented but that can be handled using a few flags.

So let's create a wrapper around irq_sources for a xive source
and provide a new API for things like PHB4 to instanciate them

This patch while at it also fixes the calculation of the source
offset when setting up targetting information in the PHB4.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>