core/pci: Support virtual device
The NVLinks (v1 and v2 to be supported in future) are exposed to
Linux kernel by emulated PCI devices (aka PCI virtual devices).
Currently, the implementation is covered by NVLink driver (npu.c),
meaning npu2.c will have similar implementation though it will be
totally duplicated with that in npu.c.

This supports PCI virtual device in the generic layer so that it
can be shared by all NVLink drivers. The design is highlighted as:

   * There are 3 config spaces for every PCI virtual device, corresponds
     to the cached config space, readonly space, write-1-clear space.
   * Reuse PCI config register filter mechanism to allow NVLink driver
     to emulate the access to the designated config registers. The config
     values are fetched from or written to the cached config space when
     the config registers aren't covered by filter.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Reviewed-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>