head.S: document opal_naca
On P7 machines OPAL can be loaded as a replacement for PHYP or in an
OPAL specific mode. In OPAL mode 0x0 contains a pointer to the OPAL
specific variant of the NACA which defines the entrypoint to OPAL, etc.
Later machines use the same load method for for PHYP and OPAL and do not
need the OPAL specific NACA, but this needs to be kept around for backwards
compatibility.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>