fast-reset: free fdt on fast reset, count fast reboots
A bit of a hack to free the flattened device tree on fast reset.
This means we don't leak ~500kb memory every fast reset.

We also count the number of fast resets we've done (if enabled),
which means that for stress testing, we have a hope of finding out
how many we managed to do before we hit a problem.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>