core/vpd: remove assert() when checking for VPD
The VPD LID is used on FSP machines to discover hotpluggable PCI slots.
Currently if we fail to load the LID skiboot will fail an assert()
preventing the system from booting. This is excessive since this is an
otherwise recoverable failure. This patch converts the assert() to an
if-NULL-return check so we can continue to boot.

Cc: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
[stewart@linux.vnet.ibm.com: fix missing verb]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>