core/vpd: remove realloc
When setting up the VPD lid preload a buffer of VPD_LID_MAX_SIZE bytes
is allocated. Once the lid has been transferred from the FSP this buffer
is shrunk with realloc. However, the contents of the buffer are copied
into the devicetree and the buffer is freed making the shrinking
pointless.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>