core/vpd: rename vpd to vpd_lid
This file contains several helper functions which also take an argument
named "vpd". Even at the best of times, the VPD parsing code is a soup
of confusing terminology so lets not make a bad situation worse. This
patch renames vpd to vpd_lid and vpd_size to vpd_lid_size to clarify the
situation.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>