hw/phb3: set PHB retry state correctly when fresetting during a creset
When we're doing a complete reset, after we complete the ETU reset and wait
for the PHB to return, we need to do a fundamental reset.

When we do the fundamental reset, we poll for a link up. This isn't always
successful on the first attempt. In phb3_sm_link_poll(), if we time out
while waiting for the link to come up, we call phb3_retry_state() to reset
p->state back to p->retry_state and poll again. On the second poll, we
clear the retry state so we don't retry again.

However, when we do the fundamental reset as part of a complete reset, we
don't explicitly set the retry state. This means that we only retry if
there wasn't an earlier fundamental reset that had to retry and thus
cleared the retry state. This reduces the reliability of the complete reset
process.

In phb3_sm_complete_reset(), when in state PHB3_STATE_CRESET_FRESET,
set the retry state to PHB3_STATE_FRESET_START, as is done in
phb3_fundamental_reset().

Reported-by: Pradipta Ghosh <pradghos@in.ibm.com>
Suggested-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Cc: Uma Krishnan <ukrishn@linux.vnet.ibm.com>
Cc: Matthew Ochs <mrochs@linux.vnet.ibm.com>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>