mambo: Make POWER9 look like DD2
In a recent Linux (next) change we added support for Radix Tree Size
(RTS) encodings for different POWER9 revs. Linux commit:

  commit 694c4951922d114e789f669deb409b2aef440ae9
  Author: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
  Date:   Wed Aug 24 15:03:37 2016 +0530
  powerpc/mm/radix: Use different RTS encoding for different POWER9 revs

In mambo, we support the DD2 encoding of RTS but the PVR we advertise
is DD1.  Hence Linux uses the wrong RTS encoding and we don't boot.

This fixes mambo to use the POWER9 DD2 PVR encoding.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>