libstb/drivers: add romcode driver
This adds a driver for the ROM verification code. The driver is compatible
with 'ibm,secureboot-v1'.

The presense of a verification code in the platform is indicated by the
presence of the ibm,secureboot node in the device tree.

The ibm,secureboot node is documented in 'doc/device-tree/ibm,secureboot.rst'

Signed-off-by: Claudio Carvalho <cclaudio@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>