mambo: Advertise available RADIX page sizes
This adds a device tree entry which advertises additional support for 2M
and 1G pages in a PAPR compliant manner. Without this, the kernel will
default to only 4K and 64K page sizes.

With this patch, 1G (0x40000000) pages for linear mapping will be used.
This can be seen by when checking the output of "Mapping range" in the
kernel log:

 # dmesg |grep "Mapping range"
 Mapping range 0x0 - 0x100000000 with 0x40000000

Without this patch, the kernel will use 64K (0x10000) instead:

 # dmesg |grep "Mapping range"
 Mapping range 0x0 - 0x100000000 with 0x10000

Signed-off-by: Chris Smart <chris@distroguy.com>
Acked-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>