core/init: Compute built-in/preloaded BOOTKERNEL size from ELF and do secure/trusted boot
Using some basic math on the ELF headers, we can work out how big the
BOOTKERNEL is that may be already loaded (e.g. in mambo, pointed to in the
device tree or built into the skiboot binary).

From there, we can measure the BOOTKERNEL in the TPM and/or verify signature.

This is only useful in a simulation environment, as real machines will
do all of this throught the load_resource API.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>