pci: Avoid hot resets at boot time
In the PCI post-fundamental reset code, a hot reset is performed at the
end.  This is causing issues at boot time as a reset signal is being sent
downstream before the links are up, which is causing issues on adapters
behind switches.  No errors result in skiboot, but the adapters are not
usable in Linux as a result.

Hot resets also occur in the FSP platform-specific code for conventional
PCI slots, which could cause issues.

This patch fixes some adapters not being configurable in Linux on some
systems.  The issue was not present in skiboot 5.2.x.

Cc: stable # 5.3.x
Signed-off-by: Russell Currey <ruscur@russell.cc>
Acked-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
(cherry picked from commit 5ac71c9b9f7a6c7dc909bdcf121f8d1d11a10dc2)
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>