core/init.c: move stb_init to later in the boot
PNOR partitions downloaded by the resource load framework are not being
measured in habanero, but they should be. The problem is that the
Nuvoton TPM driver is failing to be loaded.

The Nuvoton TPM driver is loaded in stb_init(), but when stb_init() is
called the P8 I2C bus has not been initialized yet. Error message:
"NUVOTON: ibm,opal-id property not found, tpm node parent ..."

This moves stb_init() to later in the boot to meet all its requirements.

Signed-off-by: Claudio Carvalho <cclaudio@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>