core/init.c: Fix bootargs parsing
Currently the bootargs are unconditionally deleted, which causes
a bug where the bootargs passed in by the device tree are lost.

This patch deletes bootargs only if it needs to be replaced by the NVRAM
entry.

This patch also removes KERNEL_COMMAND_LINE config option in favour of
using the NVRAM or a device tree.

Signed-off-by: Chris Smart <chris@distroguy.com>
Acked-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>