core/flash-subpartition: compute partition size from subpartition struct
From the subpartition structure, we have the ability to compute
the full partition size. Do that.

This lets us only read the amount of a subpartition that is valid
and needed to be read, rather than having to read the entire thing.

We continue the current behaviour of loading flash partitions though.

Based-on-patch-by: Claudio Carvalho <cclaudio@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>