fast-reset: disable fast reboot in event of platform error
Most of the time, if we're rebooting due to a platform error, we should
trigger a checkstop. However, if we haven't been told what we should do
to trigger a checkstop (e.g. on an FSP machine), then we should still
fail to fast-reboot.

So, disable fast-reboot in the OPAL_CEC_REBOOT2 code path
for OPAL_REBOOT_PLATFORM_ERROR reboot type.

We don't save the reason for the platform error as that's already
been error logged.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>