travis-ci: use Docker containers for a matrix of platforms
This greatly simplifies the build process for travis, yet makes it more
powerful and increases coverage without increasing wall time to test.

Travis has the concept of a build matrix, and we want to ensure we continue
to build succesfully on a variety of platforms and compiler combinations.

We limit what we run on some OSs to conserve vital sanity.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>