core/flash: Don't do anything clever for OPAL_FLASH_{READ, WRITE, ERASE}
These OPAL calls are exposing the host flash chip to linux as a flash
device. The host is expected to understand that it is raw flash and use
it appropriately, it isn't skiboots place to strip ecc bytes or perform
erase before writes.

Over the course of other developments blocklevel has gotten quite clever
but in this case the cleverness is inappropriate.

Fixes: https://github.com/open-power/skiboot/issues/44
Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>