external/common: Teach ARM code to erase 'mtd chips'
Currently the arch flash code for all architectures can only perform
chip erases if there is a real flash driver attached.

With increasing use of MTD on both host and BMC this code needs to know
how to behave of the backend of blocklevel is MTD.

This patch teaches the ARM specific code to pass an erase for the full
size of the chip down the stack. This can be optimised into a chip erase
within the kernel.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>