fsp: Don't recurse pollers in ibm_fsp_terminate
If we were to terminate in a poller, we'd call op_display() which
called pollers which hit the recursive poller warning, which ended
in not much fun at all.

This patch will skip the running of pollers and instead run
the FSP poller to set the op-panel display before attn.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>