run pollers in cpu_process_local_jobs() if running job synchonously
In the event we only have 1 CPU thread, we run asynchronous jobs
synchronously, and while we wait for them to finish, we run pollers.

However, if the jobs themselves don't call pollers (e.g. by time_wait())
then we'll end up in long periods of not running pollers at all.

To work around this, explicitly run pollers when we're the only
CPU thread (i.e. when we run the job synchronously).

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>