slw: do SLW timer testing while holding xscom lock
We add some routines that let a caller get the xscom lock once and
then do a bunch of xscoms while holding it.

In some situations without this, it could take long enough to get
the xscom lock that the 1ms timeout would expire and we'd falsely
think the SLW timer didn't work when in fact it did.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
(cherry picked from commit a5761fb4585520983716d17fdb33f04891cf0479)
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>