i2c: Add nuvoton quirk, disallowing i2cdetect as it locks TPM
In TPM 2.0 Firmware 1.3.0.1 and 1.3.1.0 (at least) there exists
a bug where if you send the wrong thing to the TPM it may lock the
bus, with no way of recovery except powering the TPM off/on.

On our current systems, the only way to power the TPM off/on is to pull
the power on the system (*NOT* just power off/on to host from BMC).

So, this patch adds the ability to do things to the i2c request really
early on, well before it hits any hardware, such as quickly drop it.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
(cherry picked from commit 157468aeb16e9f50551ac9bc7d46887cba0e011f)
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>